removed uint cast

+++ ft_strlcat.c
@@ -6,7 +6,7 @@
 /*   By: sunderle <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/08/16 17:43:03 by sunderle          #+#    #+#             */
-/*   Updated: 2020/11/24 00:24:07 by sunderle         ###   ########.fr       */
+/*   Updated: 2020/11/28 19:42:37 by sunderle         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,7 +18,7 @@ size_t	ft_strlcat(char *dest, const char *src, size_t size)
 	size_t	result;
 
 	i = ft_strlen(dest);
-	if ((unsigned int)ft_strlen(dest) < size)
+	if (ft_strlen(dest) < size)
 		result = i + ft_strlen(src);
 	else
 		result = size + ft_strlen(src);
